backend/app/ml_codes/pyproject.toml
[project]
name = "dsi-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "jupyter>=1.1.1",
    "numpy>=2.2.6",
    "ollama>=0.4.9",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.5",
    "pydantic-ai>=0.2.10",
    "requests>=2.32.3",
]
